requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.12.43
botocore==1.15.43
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
google-auth==1.14.1
idna==2.9
idna-ssl==1.1.0
jmespath==0.9.5
kubernetes==11.0.0
multidict==4.7.5
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
slackbot==0.5.6
slackclient==2.5.0
slacker==0.14.0
typing-extensions==3.7.4.2
urllib3==1.25.9
websocket-client==0.44.0
yarl==1.4.2
